+++ src/compas_viewer/ui/mainwindow.py
