+++ src/compas_viewer/ui/statusbar.py
